Added HelloNano
This was basically just a simple application for Atmel Studio 7, generated for the Atmel4809 Curiosity Nano with the Atmel Start web application, just to see if I could get it to work. There is no board support file in Atmel Start but it looks like one is not really needed. Just keep in mind that the USB serial port is USART3.

The program repeatedly prints "Hello, Nano" to the terminal that you can open with Tools > Data Visualizer. Make sure the bit rate is set to 200,000 bps.