Committing first verison of deck control analyzer
The DDU2113 DCC deck is used in the DCC-951, DCC-730 and the FW-68. It communicates with the microcontroller on the digital print via a simple two-wire serial port connection. A quick analysis with an oscilloscope showed that the protocol is 38400 8n1. Async serial is easy to understand, so I wrote this as the first analysis application just to see if I could get it to work and if I could analyze the data that goes back and forth.

This program was generated with Atmel Start, the only code I changed is in main.c. I used an ATMega4809 Curiosity Nano, for which there is no board support file in Atmel Start. But it doesn't really need one; all you need to know is that the serial port that's connected to the USB debugger is USART3.

In this app I configured USART3 to run at 200,000bps 8n1, and configured USART0 and USART1 in polling receive-only mode at 38400bps. It may not be the most efficient, but it works and it's fast enough to catch all the data and print it to the terminal.

The DIG MCU issues a command to the DDS MCU every 35ms approximately. Most of the time, this is just a status update request but if the user pushes a button on the front panel or sends a remote control command, it sends a different command and the reply is usually shorter. This first commit just dumps the command and the response to the terminal in hexadecimal. See the remarks in the code for what I found out about the DIG-DDS communication so far.

This generates a lot of data on the terminal; I'll work on that next.