Deck Control Monitor program working (mostly)
Updated the deck control monitoring program to interpret the data and only show text when something changes.

I had to change the structure of the program so that it adds the information about the command and response to a string buffer as it goes, and decides whether something has changed since last time. Whenever there is no incoming traffic from either of the ports that are on the recorder, the code sends a byte from the output buffer to the computer. All of this is done very crudely with printf's which is probably horribly and unnecessarily inefficient. And because there's only one output buffer, it sometimes doesn't get all of it out to the serial port. That should be easy to fix by using a circular buffer and not discarding the output buffer when a new command is issued by the DIG MCU.

All in all a nice exercise and now I know there are a few commands that it apparently never uses. I wonder what they would do :-)