Re-created the app from scratch in Atmel Start
Apparently Atmel Start is too braindead to accept that this is now my app so it kept renaming the application back to "EDBG UART" and kept insisting on writing the old main file back.
This also appears to have changed some clock settings to other defaults but as far as I can tell, that's not really important.
I saved the Atmel Start project file too so it's possible to recreate this from scratch, from that single file and the modified main.c