Added SPI_EXT2 and switched both SPI controllers to LSB first
I think the NEC/Renesas datasheet for the 78K microcontrollers that Philips used in their DCC recorders labels the synchronous serial connections exactly like that: Synchronous serial. The only difference between the protocol in the DCC recorder and the now commonly used TTL serial ports is that the bytes don't have a start bit. So the clock line is high when idle, goes low at the start of a transfer, the (8) bits are sampled when the clock goes high, and bits are transferred LSB first. Though none of my microcontrollers can do that with their versions of synchronous serial (UARTS insist on at least one start bit in sync serial mode), it's easily compatible with SPI mode 3, MSB first.