Added LED output, changed debug console to 115200bps
It looks like this works but not if you turn the recorder on while the program is already running.
I speculate that the shift registers of the SPI ports go out of sync because when the recorder powers on, the SCLK line goes from low to high (maybe multiple times) and the microcontroller thinks that that's a valid clock pulse and samples the bit.
I tried putting a pull-up resistor on the SCLK lines but that had no effect other than the program not registering any data after a few seconds.
It may be possible to solve this by using MESSYNC. I'll look into it later. For now the workaround is to reset the microcontroller while the recorder is already powered up. Until then, we won't be able to see the data exchange at initialization time.