Added code to let the button restart the SPI controllers
I got fed up with the application going out of sync (bitwise) when the recorder was started after my MCU was already running, so I added some code to let the button on the Xplained Pro board disable and re-enable the SPI buses.

That seemed to do the trick realy nicely but I decided to also check the MESSYNC signal on the logic analyzer. Turns out it goes high between messages (and also when HOLD is high) so I connected the -SS of both SPI buses to that signal and now the button isn't needed anymore.

Also added some code to check for bad checksums. They are indicated with "<!>"