Added some extra buffers
Added some code to transfer the incoming bytes out of the SPI instances' ring buffers which are not adjustable and probably too short (I expect a single message in our DCC recorder may be at least 42 bytes).
Also changed the code that waits for each pair of (command/response) bytes so it uses the ringbuffer_num function. Still a lot of busy-waiting but now at least it looks like we're not dropping any bytes anymore, and all the checksums apparently are correct. Yay!
Also modified the Atmel Start project so that the first two GPIO lines to the debugger are now available. I did this while I was going through 4 or 5 revisions to try and read all the bytes from the callbacks. That code was not checked in because I couldn't get it to work. Too many race conditions (surprisingly).