Reverse-engineered front panel bus
Not everything has been analyzed yet, this is a work in progress to give me some idea of what the recorder can do.
Some commands are muted because they are very chatty (e.g. VU updates). It's obvious that some of those commands are used to poll for updates (e.g. check for markers on the tape). I'll have to look into those further.