Removed stdio redirection and substituted my own
Atmel Start doesn't let you configure stdio redirection with a SPI port, and even with the workaround in the previous commits (where I just had it connected to something else in Start but changed it to the EDBG SPI output at startup time), it doesn't activate and deactivate the Slave Select line for the EDBG slave (I previously just set the pin to Low at beginning of world and never touched it). This is a problem if other devices are connected to the same SPI port, such as a display.