Debug output now goes to EDBG serial port
Also removed the EXT3 SPI port which is connected to the debug chip.
I may re-add it later.

To see the output of the front panel monitor, connect a terminal emulator to the EDBG Virtual COM port and configure it to 1000000 (one million) bits per second.

Note: Actually I re-created the project for this commit: I loaded the .atstart file into the Atmel Start website at start.atmel.com and then saved it. For some reason it was the only way I could make the stdio redirection "stick". I suspect that I had stdio redirection in the project before and then removed it manually instead of by using the Start tool, and it didn't update the files correctly. Anyhow,  it looks like stdio redirection to the EDBG serial port is working now so it's not necessary to set up the terminal to the SPI port in the Data Visualizer anymore.