Added L3 monitoring to the front panel monitor
The following connections were added:
EXT3_MOSI=L3DATA
EXT3_SCK=L3CLK
EXT3_IRQ=L3MODE_IRQ
EXT3_SPI_SSA=(GND)

Note: The SPI port on the EXT3 connector that I used for this, is also connected to the debug chip.

There are two callback functions:
- l3mode_callback is called whenever the L3MODE input changes from LOW to HIGH or vice versa
- l3spi_callback is called whenever a character arrives at the MOSI input of SERCOM7.

Whenever L3MODE changes, the global variable l3mode is changed to let the incoming data callback know that the data is an address (L3MODE = LOW) or data (L3MODE = HIGH). If L3MODE goes from HIGH to LOW, the mode change callback marks the current buf_t buffer as ready and moves to the next buffer in the array of buffers. If the next buffer is not free, the overrun counter is increased.

The incoming data callback stores the data in the current buffer (unless it's already marked "ready" indicating a buffer overrun). If the buffer doesn't have the response index (really the index of the start of data) set, it sets the index.

The main routine was changed so it now simply calls the two data capturing functions: one for the front panel and one for the L3 bus. All massaging (e.g. getting data from the ring buffers and clearing msb's of the command codes in the front panel) is done in the capture functions, no longer in the main function.

The L3 capture contains some commented-out code based on the SAA2023 datasheet but I suspect it's all wrong. I'll look into how the data is supposed to be correctly interpreted, next.