Initial commit of the Windows DCC730/951 Front Panel Monitor
This will use two FTDI UMFT4222EV-D boards and a 74LVC245 to spy on the front panel to deck connection using Windows. It will use knowledge gained from the Software/SAMC21N-XPRO/FrontPanelMon project and will possibly reuse some code.

Later on, this project will be ported to the Teensy 4.1, but at this time I don't have a fully working SPI Slave stack for that and I think it would be too big a step.