Added minimal receive and hex dump code
The code works with a modified version of the no protocol SPI master program from the FT4222 samples (the code must be modified to always send data, never receive data. Otherwise the transmitter will keep waiting for a sync byte that never comes).

The user can exit the receiver by hitting Q on the keyboard.