Added code to receive using SPI mode 3
The DCC hardware uses serial ports with a clock. The data line changes when the clock goes low (so IDLE_HIGH) and the data signal should be read on the trailing edge of the clock (so CLK_LEADING).

The bits are transferred LSB to MSB but unfortunately that can't be accomplished using FT4222 hardware.